MessageType支持byte,short,int三种，用于支持不同数量的MessageType

+++ easysocket-netty-msg/src/test/java/cn/featherfly/network/netty/msg/NettyMessageClientTest.java
@@ -17,10 +17,10 @@ public class NettyMessageClientTest {
     static MessageTypeRegister register;
     static {
         register = new MessageTypeRegister();
-        register.register(TestClientMsg.class, (short) 0);
-        register.register(ClientRegistMsg.class, (short) 1);
-        register.register(ResponseMsg.class, (short) 2);
-        register.register(BasicMsg.class, (short) 3);
+        register.register(TestClientMsg.class, 0);
+        register.register(ClientRegistMsg.class, 1);
+        register.register(ResponseMsg.class, 2);
+        register.register(BasicMsg.class, 3);
     }
 
     public static void main(String[] args) {
